moved collisioncontroller to gamestatecontroller
